Fixed failed test and reduced code duplications
