+++ .sonarcloud.properties
@@ -1,1 +1,7 @@
 
+sonar.sources=.
+sonar.exclusions=examples/**,configs/testdata/**,**/*_test.go,**/vendor/**,internal/pkg/data/*datastore.go
+
+sonar.tests=.
+sonar.test.inclusions=**/*_test.go
+sonar.test.exclusions=**/vendor/**
